fix(cms): fix ui component import
